refactor: Add some unit-tests and refactor the code to make it more readable
