+++ hello/tests/test-requirements.txt
@@ -1,0 +1,3 @@
+boto3
+moto
+pylint
